requirements.txt
loguru
numpy
Pillow
PyOpenGL
pyrender
trimesh
requests
httpx
openai
pydantic
PyYAML
boto3
python-dotenv
tqdm